Move import views/viewModels to separate namespace
