+++ WPFSKillTree/Views/Import/ImportCharacterWindow.xaml
@@ -1,5 +1,5 @@
 ﻿<dialogs:CloseableBaseDialog
-    x:Class="PoESkillTree.Views.Equipment.ImportCharacterWindow"
+    x:Class="PoESkillTree.Views.Import.ImportCharacterWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:l="clr-namespace:PoESkillTree.Localization.XAML"
@@ -7,7 +7,7 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d"
     xmlns:metroControls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
     xmlns:dialogs="clr-namespace:PoESkillTree.Controls.Dialogs"
-    xmlns:viewModels="clr-namespace:PoESkillTree.ViewModels.Equipment"
+    xmlns:viewModels="clr-namespace:PoESkillTree.ViewModels.Import"
     d:DataContext="{d:DesignInstance viewModels:ImportCharacterViewModel}">
     <StackPanel>
         <TextBlock Margin="5" TextWrapping="Wrap">
