+++ WPFSKillTree/Views/Import/ImportStashWindow.xaml.cs
@@ -1,4 +1,4 @@
-﻿namespace PoESkillTree.Views.Equipment
+﻿namespace PoESkillTree.Views.Import
 {
     /// <summary>
     /// Interaction logic for ImportStashWindow.xaml
